soft-delete, validation, multi-threading

+++ MyCart/src/main/java/com/example/mycart/controller/AbstractGenericController.java
@@ -1,9 +1,12 @@
 package com.example.mycart.controller;
 
+import com.example.mycart.exception.ApiException;
+import com.example.mycart.exception.ResourceNotFoundException;
 import com.example.mycart.model.BaseEntity;
 import com.example.mycart.payloads.BaseDTO;
 import com.example.mycart.service.GenericService;
 import com.example.mycart.modelmapper.EntityMapper;
+import com.example.mycart.utils.Validator;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
@@ -36,12 +39,19 @@ public abstract class AbstractGenericController<T extends BaseEntity<Long>,D ext
     {
         var result = getService().findById(id);
 
+        if(result==null)
+            throw new ApiException("unable to find request object with id: "+ id);
+
         return new ResponseEntity<>(getMapper().toDTO(result,pageNo), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<D> delete(@PathVariable ID id, @RequestParam(defaultValue = "0", required = false) int pageNo)
     {
+        if(new Validator().validateEntity(getService(),id))
+        {
+            throw new ApiException("unable to find request object with id: "+ id);
+        }
         var result = getService().delete(id);
 
         return new ResponseEntity<>(getMapper().toDTO(result,pageNo), HttpStatus.OK);
@@ -50,6 +60,11 @@ public abstract class AbstractGenericController<T extends BaseEntity<Long>,D ext
     @PutMapping("/{id}")
     public ResponseEntity<D> update(@PathVariable ID id,@Valid @RequestBody D dto, @RequestParam(defaultValue = "0", required = false) int pageNo)
     {
+        if(new Validator().validateEntity(getService(),id))
+        {
+            throw new ApiException("unable to find request object with id: "+ id);
+        }
+
         var result = getService().update(id,dto);
 
         return new ResponseEntity<>(getMapper().toDTO(result,pageNo), HttpStatus.OK);
