+++ MyCart/src/main/java/com/example/mycart/controller/CartController.java
@@ -1,5 +1,6 @@
 package com.example.mycart.controller;
 
+import com.example.mycart.exception.ResourceNotFoundException;
 import com.example.mycart.model.Cart;
 import com.example.mycart.model.CartItem;
 import com.example.mycart.modelmapper.CartItemMapper;
@@ -7,6 +8,7 @@ import com.example.mycart.modelmapper.CartMapper;
 import com.example.mycart.payloads.CartDTO;
 import com.example.mycart.payloads.CartItemDTO;
 import com.example.mycart.service.CartService;
+import com.example.mycart.utils.Validator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -27,9 +29,17 @@ public class CartController
     @Autowired
     private CartItemMapper<CartItem,CartItemDTO> cartItemMapper;
 
+    @Autowired
+    private Validator validator;
+
     @GetMapping("/user/{userId}")
     public ResponseEntity<CartDTO> getCart(@PathVariable Long userId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
+
         var cartItems = service.findCartByUser(userId);
 
         return new ResponseEntity<>(cartMapper.toDTO(cartItems,0), HttpStatus.OK);
@@ -41,6 +51,16 @@ public class CartController
             @PathVariable Long userId,
             @PathVariable Long productId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
+
+        if(validator.validateProduct(productId))
+        {
+            throw new ResourceNotFoundException("Product","id",productId);
+        }
+
         var cartItem = service.addItemToCart(userId, productId, cartItemDTO.getQuantity());
 
         return new ResponseEntity<>(cartItemMapper.toDTO(cartItem,0),HttpStatus.CREATED);
@@ -52,6 +72,10 @@ public class CartController
             @PathVariable Long cartItemId,
             @PathVariable Long userId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
         var cartItem = service.updateCartItemQuantity(userId,cartItemId, cartItemDTO.getQuantity());
 
         return new ResponseEntity<>(cartItemMapper.toDTO(cartItem,0),HttpStatus.CREATED);
@@ -61,6 +85,11 @@ public class CartController
     public ResponseEntity<CartItemDTO> removeItemFromCart(@PathVariable Long cartItemId,
                                                           @PathVariable Long userId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
+
         var deletedItem = service.removeItemFromCart(userId,cartItemId);
 
         return new ResponseEntity<>(cartItemMapper.toDTO(deletedItem,0),HttpStatus.OK);
@@ -69,6 +98,10 @@ public class CartController
     @DeleteMapping("/user/{userId}")
     public ResponseEntity<CartDTO> clearCart(@PathVariable Long userId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
 
         var cart = service.clearCart(userId);
