+++ MyCart/src/main/java/com/example/mycart/controller/OrderController.java
@@ -1,10 +1,12 @@
 package com.example.mycart.controller;
 
+import com.example.mycart.exception.ResourceNotFoundException;
 import com.example.mycart.model.Order;
 import com.example.mycart.payloads.OrderDTO;
 import com.example.mycart.service.OrderService;
 import com.example.mycart.modelmapper.EntityMapper;
 import com.example.mycart.utils.OrderStatus;
+import com.example.mycart.utils.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -25,9 +27,17 @@ public class OrderController
     @Autowired
     private EntityMapper<Order,OrderDTO> orderMapper;
 
+    @Autowired
+    private Validator validator;
+
     @PostMapping("/user/{userId}")
     public ResponseEntity<OrderDTO> createOrder(@PathVariable Long userId)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
+
         var order = service.create(userId);
 
         return new ResponseEntity<>(orderMapper.toDTO(order,0), HttpStatus.CREATED);
@@ -38,12 +48,21 @@ public class OrderController
     {
         var order = service.findById(orderId);
 
+        if(order==null)
+        {
+            throw new ResourceNotFoundException("Order","id",orderId);
+        }
         return new ResponseEntity<>(orderMapper.toDTO(order,0),HttpStatus.OK);
     }
 
     @GetMapping("/user/{userId}")
     public ResponseEntity<Page<OrderDTO>> getUserOrders(@PathVariable Long userId, @RequestParam(defaultValue = "0",required = false) int pageNo)
     {
+        if(validator.validateUser(userId))
+        {
+            throw new ResourceNotFoundException("User","id",userId);
+        }
+
         var orders = service.getOrdersByUser(userId, pageNo);
 
         return new ResponseEntity<>(orders.map(order -> orderMapper.toDTO(order,0)),HttpStatus.OK);
@@ -54,6 +73,11 @@ public class OrderController
             @PathVariable Long orderId,
             @PathVariable OrderStatus value)
     {
+        if(validator.validateOrder(orderId))
+        {
+            throw new ResourceNotFoundException("Order","id",orderId);
+        }
+
         var updatedOrder = service.updateOrderStatus(orderId, value);
 
         return new ResponseEntity<>(orderMapper.toDTO(updatedOrder,0),HttpStatus.OK);
@@ -62,6 +86,11 @@ public class OrderController
     @DeleteMapping("/{orderId}")
     public ResponseEntity<OrderDTO> cancelOrder(@PathVariable Long orderId)
     {
+        if(validator.validateOrder(orderId))
+        {
+            throw new ResourceNotFoundException("Order","id",orderId);
+        }
+
         var cancelledOrder = service.cancelOrder(orderId);
 
         return new ResponseEntity<>(orderMapper.toDTO(cancelledOrder,0),HttpStatus.OK);
@@ -70,6 +99,10 @@ public class OrderController
     @DeleteMapping("/{orderId}/product/{productId}")
     public ResponseEntity<OrderDTO> cancelOrderItem(@PathVariable Long orderId, @PathVariable Long productId)
     {
+        if(validator.validateOrder(orderId) && validator.validateProduct(productId))
+        {
+            throw new ResourceNotFoundException("Order OR product","id",orderId);
+        }
         var cancelledOrder = service.removeOrderItem(orderId,productId);
 
         return new ResponseEntity<>(orderMapper.toDTO(cancelledOrder,0),HttpStatus.OK);
